Add files via upload
The generated report adapts the total sales price and taxes in the entered vehicle galleries according to the records given in the database table. There are two buttons called "Round prices" and "Maximize prices". The round price button brings car prices in showrooms to the lowest possible level. The button to maximize vehicle sales optimizes sales by reducing the profit of the dealer in the showroom. (I use my own alv display structure)